requirements.txt
Django==2.2.3
django-widget-tweaks==1.4.1
djangorestframework==3.10.2
python-memcached==1.59
